refactor hex() and oct().
In py2 they support the special methods __hex__ and __oct__, and thus are
rendered as regular objspace methods (space.oct and space.hex).

In py3k, __hex__ and __oct__ are gone, so they are just normal builtins, which
are implemented on top of __format__ for consistency with bin() (the
alternative would be to implement them on top of the rpython hex() and oct()).

Killing support for hex and oct from the objspace is not trivial, e.g. we need
to keep them in the method table because the flow objspace expects them. I
hope not to have broken anything. Any review is appreciated :-)

--HG--
branch : py3k